Fix package files list + Fix backup files
The list containing the files for tar to assemble was created with some
files beginning at a new line while printing other just right next to
each other with just a space in between. This led to any important
metadata files not being present in the package.
Add backup files to the PKGINFO.
Escape bash variables possibly containing spaces.
State that reassembling may take a while in the man page and point out
how to speed things up a little.